Dramatically improve error messages with source query pointers (#7335)
## Release notes: product changes

TypeDB now shows the source of an error in the context of the original
query, wher it is possible. In general, we aim to show a detailed error
message in context of the original query whenever the error arises in
the compilation phase of the query.

Example of the improved error format:
```
[QEX2] Failed to execute define query.
Near 4:29
-----
        define
        attribute name value string;
-->     entity person owns name @range(0..10);
                                ^
        
-----
Caused by: 
      [DEX25] Defining annotation failed for type 'person'.
Caused by: 
      [COW4] Concept write failed due to a schema validation error.
Caused by: 
      [SVL34] Invalid arguments for range annotation '@range(0..10)' for value type 'Some(String)'.
```

## Motivation

TypeDB 3.x's error messages show a "stack trace" of causation that lead
to a particular failure mode. However, these errors almost always
originate somewhere in the user's original query. TypeDB now includes a
snippet of the original query, pinpointing (to best-effort) the place in
the query that triggered the failure. This improves self-help and
debuggability dramatically.

## Implementation

* We add `source_query` and `source_span` fields as special entries into
`TypeDBError` macros invocations. Error stacks that contain both a
source query and a source span, will on formatting produce an excerpt of
the query based on the source span.
* We add source spans throughout the `//ir` layer, which will be used to
produce errors if they are encountered
* In general, we do a best-effort placement of the Span, based on the
most-accurate possible origin in the TypeQL query

### Tradeoffs
We do introduce some architectural awkwardness: all packages that use
`TypeDBError` macros must depend on `TypeQL`, in case they decide to
include a `Span` query pointer. This includes low-level system packages!

We choose to accept this instead of building an alternative
representation of `Span` that isn't tied to TypeQL to avoid major
complexity: converting a `Span`, which is what TypeQL currently offers,
to a line, column `usize` pair would require access to the query string.
The 'correct' solution may be to change TypeQL to pre-compute (line,
col) positions as well as/instead of Spans, that TypeDB can use
throughout.