Add flag to type-seeder for write stages (#7338)
## Release notes: product changes
Adds a flag to type-seeder to indicate the stage is a write stage. This ensures variables constrained by `isa`, or labelled roles are seeded with the exact type, and not (transitive) subtypes. This fixes a bug where one could not insert a role-player for a role with a subtype.

## Motivation
Fixes #7333 

## Implementation
Adds a flag to type-seeder to indicate the stage is a write stage. This ensures variables constrained by `isa`, or labelled roles are seeded with the exact type, and not (transitive) subtypes.
Updates the TypeSeeder `UnaryConstraint` for Isa and RoleName