Avoid persisting empty statistics deltas (#7347)
## Release notes: product changes

We resolve the issue in which statistics would be written to the WAL
more often than necessary, which could cause WAL (not the data) to
become corrupted.

## Motivation

If a schema commit was performed within 8 commits before the statistics
was last updated, the sequence number stamp in the statistics would roll
back, but not the counts. Any intermittent commit deltas would be
reapplied repeastedly, in some cases causing underflows and crashes.

In addition, statistics would be unconditionally persisted to the WAL
when said schema commit was encountered, adding needless repeated
records _and_ persisting the corrupted statistics records, causing
startup crashes.

## Implementation

We now keep track of when the last write was performed, as well as fix
the issue with rolling back the sequence number.