Improve query plan eviction (#7366)
## Release notes: product changes

We improve the query plan eviction policy from a flat 1% change in total
count of the data, to instead be a 25% change of _any individual_
statistic, such as a specific entity type's instances, or the count of
`has` between two specific types, etc. This means the query plans will
be much more accurate in rapidly changing data, in particular in the
common case of building an initial dataset or benchmarking.

Before, we'd get tpcc results for the PAYMENT action that look like
this:
```

Execution Results after 63 seconds
--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
                  Complete        Time (µs)         minLatMs        p50             p75             p90             p95             p99             maxLatMs        Aborts                   
...
  PAYMENT         20                 61.618         7.90            12.91           1244.86         15993.48        24360.62        24360.62        24360.62        0                    
...
--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
```

This shows that during a 60s execution time, a few pathological query
executions would blow up the entire runtime! However, sometimes the
query also executed quickly. This was fundamentally caused by the
initial queries using the query plan based on old statistics, which
would only get updated later in the benchmark when the total statistics
updated sufficiently.

## Motivation

We move towards the goal of being able to reliably execute TPCC
benchmarks without needing to force query plans to replan manually.

We now reliably get TPCC results for PAYMENT that look like this
instead:

```
Execution Results after 60 seconds
--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
                  Complete        Time (µs)         minLatMs        p50             p75             p90             p95             p99             maxLatMs        Aborts                
...
  PAYMENT         428                 43.54         7.00            14.05           26.01          184.51          249.36          281.22          675.73          0               
...
--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
```


## Implementation

The query plan cache now retains the last full set of statistics it was
flushed against, which allows calculating a much finer diff. We then
take the maximum percentage that any individual statistic has changed as
the signal of whether to flush the cache. For now, we increase that
threshold to 25% of the minum of the old or new statistics, whichever is
smaller.