Introduce update queries (#7361)
## Release notes: product changes
We introduce update queries as a shortcut for `delete + insert`. Using
`update`, it is possible to replace an existing `has` or `links` edge of
a specific attribute or role type by a new value in one step.

If an old value exists, it is deleted. Then, the new value specified in
the query is set.

This feature is only available for `has` and `links`. Moreover, to make
sure that no excessive data is unexpectedly deleted, it only works with
types with `owns` and `relates` of cardinalities not exceeding 1:
`card(0..1)` (the default value) or `card(1..1)` (card of `@key`).

Usage examples:

```
match
  $p isa person, has name "Bob";
update
  $p has balance 1234.56789dec;
```

```
match
  $alice isa person, has name "Alice";
  $bob isa person, has name "Bob";
  $e isa employment, links (employer: $alice);
update
  $e links (employee: $bob);
```

## Motivation
We want to introduce a shorter query for faster data updates instead of
using `delete` and `insert`.

## Implementation
We extend IR, annotation, compilation, and execution with a new stage
and a new set of instructions. As update queries behave similarly to
insert (they insert new data), many insert-related methods are reused,
with additional logic and validation on top. More in the PR comments.