Match disjunction variables as optional (#7356)
## Release notes: product changes

We treat variables that are only used in a branch of a disjunction as
optionally matched. If the branch is not used by a given row, the column
is left empty.

## Motivation

## Implementation

All named variables in a nested pattern are considered to be visible
from every other subpattern. Previously, any variable referenced in a
subpattern of a conjunction (viz. in a disjunction branch) would only be
an output if it was also referenced _outside_ of the subpattern.

Negations are the exception: if a variable is only introduced and used
in a negation, its value does not escape the negation's scope. If a
value were to be successfully bound to this variable, the negation
constraint would be violated, and the value not returned in the first
place.