Allow abstract relation types having zero role types. (#7378)
## Release notes: product changes
Allow abstract relation types having zero role types. It is no longer a
requirement for a relation type to declare or inherit role types if it's
abstract.

Non-abstract relation types are still required to have at least one role
type.
Notice that it can be a single abstract role type. However, such a
schema does not allow preserving any data related to this relation type
as its instances will get cleaned up on commits due to the absence of
role players.

Additionally, a bug in relation type deletion has been fixed. In some
cases, deletion of a relation type could affect specialized role types
of its supertype.

## Motivation
This change makes schemas less restrictive and more flexible.

## Implementation
### Validations
Commit time validation:
* ✓ If a relation type is not abstract, but does not have any role types
(abstract or not abstract, whatever), we error

Operation time validation (if not implemented, explained why):
* ✘ New relation type created: it can have 0 role types until we commit
it
* ✘ Relation supertype changes to a type without role types: it could
just be created, and it would be strange to error. It is the
responsibility of the supertype to have at least one role (which will be
inherited)
* ✘ Role type gets deleted: I initially wanted to check that it's not
the last role type of a type. However, it blocks the migration of form
`undefine relates rol from rel; rel;` <- it would not let to delete the
last role `rol`, which is logically executed first. This query feels
reasonable for "total cleanup of types". Maybe, if we want to keep it in
sync with the lower checks, we should check it.
* ✓ Abstractness is unset for a relation type without role types: at
least something. The error would be a little strange in case of this
series of queries: `define relation rel; ... define rel @abstract; ...
undefine @abstract from rel`, but we already have similar cases in other
constraints checks, if I remember correctly, so it's better to at least
limit this action. The only issue here is `undefine @abstract from rel;
rel;` <- same as for role types, but the intent of this query feels
different, so I'd error.
* ✓ Supertype is unset (`undefine sub rel2 from rel`), and there are no
more role types declared or inherited, while the subtype is not
abstract. Same as before: it feels reasonable, but `undefine sub rel2
from rel; rel;` <- will get a little unreasonable error.

### Relation type deletion and its effect on role types
While I thought that I misused specialized relates only in undefine
queries, it appeared that we had a bug in TypeManager. When deleting a
relation type, it cleaned up its role types, but it used to get these
role types based on declared relates... Which include both explicit and
implicit relates. Now it's fixed, and an additional Concept API
interface for Concept BDD (and other possible use) has been introduced,
similarly to name-based search methods for relates.