Fix erroneously selected named variables (#7375)
## Release notes: product changes

We adjust the computation of bound variables after each stage in query
execution and keep better track of selected variables.

## Motivation

Before this change, all named variables referenced in a match block
would be marked as being output at the end of the stage. That was
incorrect and led to anything from extra empty columns in the output
rows to spurious crashes.

## Implementation

We introduce computation of a pattern's dependency on a variable: when a
pattern is referencing a variable, it may be in a way that:
- requires it to be bound by another pattern (`Required`),
- may bind the variable to a new value if previously unbound
(`Producing`),
- neither requires it to be bound nor may bind it (`Referencing`).

An example of the latter is a disjunction that references a variable in
just one branch: `{ $x has $a; } or { $y has $a; };`. This disjunction
can produce `$a`, but only references `$x` and `$y` which may either be
fully internal to either branch or bound in the enclosing conjunction.

Only input variables and named variables marked as produced by the root
conjunction of a match are selected at the end of the block. (NB:
anonymous variables may be selected by an anonymous function in fetch).

To simplify the planning portion, we also revert disjunctions optionally
binding variables referenced in a subset of branches.