Introduce branch provenance tracking (#7402)
## Release notes: product changes
Introduce branch provenance tracking. This allows us to determine which
branch of a disjunction a given answer originates from. Though
meaningless in the logical semantics, this is helpful for other
applications like visualisations.

## Motivation
Provenance & query-structure tracking for visualisations.  

## Implementation
- Introduces `branch_id` in IR disjunctions
- Introduces one `u64` per row for branch provenance tracking.
- `DisjunctionExecutor::map_output` sets the bit corresponding to the
`branch_id` on the row's provenance.
- Introduces the `QueryStructure` structure with the information to
reconstruct the query IR from the cached executable pipeline - so we may
send it to the client side.