HTTP write queries return some answers on limit & more tests (#7437)
## Release notes: product changes
When the query option `answerCountLimit` is hit in write queries, the
transaction service no longer returns errors. Instead, the whole query
is executed, but only a limited number of answers is returned with a
warning string, similar to read queries.

Additionally, more behavior tests for HTTP-specific features are
introduced.
 
## Motivation
`answerCountLimit` is a network-based limitation, thus it should not
affect the actual execution, limiting the users to load big batches of
data.

## Implementation