Execute deletes in two passes - first constraints, then concepts (#7442)
## Release notes: product changes
If a delete stage which deletes both constraints and concepts, a certain
row may delete a concept referenced by a connection in a subsequent row.
We avoid this problem by first deleting constraints for every row in an
initial pass, followed by concepts of every row in a second pass.