Implement schema retrieval (#7424)
## Product change and motivation

We implement the (pre-existing) endpoint to retrieve the schema in two
variations. These used to return an 'unimplemented' error.

- get the define query that specificies the Types (on the driver, this
is available via a retrieved database's `database.type_schema()`)
- get the define query that specifies the Types and the Functions (on
the driver, this is available via a retrieved database's
`database.schema()`)

This API is useful to export and save an existing database schema, and
also in particular useful for Agentic systems which use the schema
definition to generate read or write queries for a database.

## Implementation

We implement the schema retrieval endpoint for the database service,
down to the Database and the TypeManager. This required a new trait,
`TypeQLSyntax`, which is implemented by each `Kind`, and various other
internal structs that must be converted into TypeQL syntax on export.

Each part of the exported schema (types, capabilities, and annotations)
is written sorted, so it is deterministically produced and tested.