Add 'z' parsing for datetime-tz (#7460)
## Product change and motivation

We fix a bug where TypeDB server failed to recognise "Z" as a valid
timezone specifier. According to ISO 8601
(https://en.wikipedia.org/wiki/ISO_8601#Coordinated_Universal_Time_(UTC)),
this is valid timezone specifier equal to the 0 offset. The underlying
time-zone library the server utilised did not accept it as a valid
timezone.

Note that TypeQL already accepted 'z' as a valid timezone specifier in
its grammar.

## Implementation

We check for the 'z' timezone explicitly, and set the fixed timezone to
the `0` fixed offset explicitly.