Add port 8000 to list of ports on docker image (#7472)
## Product change and motivation
Add port 8000 to the docker image, so the HTTP API is exposed. To be able to access it from the host machine, `-p 1729:1729 -p 8000:8000` must still be specified in the `docker create` or `docker run` commands.

## Implementation
This does indeed "expose" the ports, which can be seen using `docker inspect`. So it's more complete, though it can't remove the need to "publish" the ports at container creation.