Update planner cost model for joins to reflect seeks being implemented (#7478)
## Product change and motivation
Updates the cost for a join between iterators to reflect seeks being
introduced (#7367). The sum of the iterator sizes was accurate when we
had to scan through the iterator. With seeks, we can skip impossible
values. The number of seeks is on the order of the size of the smaller
iterator.

---------

Co-authored-by: Dmitrii Ubskii <18616863+dmitrii-ubskii@users.noreply.github.com>