Open-up server state component for extension (#7479)
## Product change and motivation

The `ServerState` component has been converted into a trait, with the
actual implementation moved into a new component `LocalServerState`.
This change is motivated by the need of allowing additional server state
implementations with extended functionalities.

## Implementation
- Introduced a server state `ServerState` trait its boxed-type companion
`BoxServerState`. It is now possible to create additional server state
component with more functionalities by implementing this trait.
- Renamed the existing `ServerState` struct to `LocalServerState`, as it
is a server state component responsible specifically for managing the
state of a single server instance.
- Various formatting cleanups