Commit statistics to WAL more frequently (#7480)
## Product change and motivation

Rebooting a large database can take a long time, due to having to
synchronise statistics from the last durably written checkpoint in the
WAL. This statistics record was written on a _percentage_ change of the
statistics, which meant a lot of transactions could be committed before
the statistics are checkpointed again. These all needed to be replayed
to generate the newest statistics values on reboot.

Instead, each database now checkpoints the statistics either every 1k
commits, or after am aggregate delta of 10k counts in the statistics,
whichever comes first. This helps bound the amount of WAL entries
replayed (& therefore memory required) for synchronising statistics on
reboot.

Note: There is still a gap in this implementation, which occurs when
there are 999 large commits happen but the aggregate delta is 0 (for
example, by adding/deleting data continuously in large blocks) - this
scenario would still require loading a large amount of the WAL into
memory on replay.

## Implementation

* We checkpoint statistics every 1k commit (determined by sequence
number) or a aggregate delta of 10k (determined by total count vs last
committed total count)