Introduce file-based database export and import (#7477)
## Product change and motivation
Add database export and database import operations. Unlike in TypeDB
2.x, these operations are run through TypeDB **GRPC clients** such as
the **Rust driver** or **TypeDB Console**, which solves a number of
issues with networking and encryption, especially relevant for users of
TypeDB Enterprise. With this, it becomes an official part of the [TypeDB
GPRC protocol](https://github.com/typedb/typedb-protocol/pull/224).
Both operations are performed through the network, but the server and
the client can be used on the same host.

Each TypeDB database can be represented as two files:
1. A text file with its TypeQL schema description: a complete `define`
query for the whole schema.
2. A binary file with its data.

This format is an extension to the TypeDB 2.x's export format. See more
details below for version compatibility information.

### Database export
Database export allows a client to download database schema and data
files from a TypeDB server for future import to the same or higher
TypeDB version.

The files are created on the client side. While the database data is
being exported, parallel queries are allowed, but none of them will
affect the exported data thanks to TypeDB's transactionality.
However, the database will not be available for such operations as
deletion.

**Exported TypeDB 3.x databases cannot be imported into servers of older
versions.**

### Database import
Database import allows a client to upload previously exported database
schema and data files to a TypeDB server. It is possible to assign any
new name to the imported database.

While the database data is being imported, it is not recommended to
perform parallel operations against it.
Interfering actions may lead to import errors or database corruption.

**Import supports all exported TypeDB 2.x and TypeDB 3.x databases.** It
can be used to migrate between TypeDB versions with breaking changes.
Please visit [our docs](https://typedb.com/docs/manual/migration/2_to_3)
for more information.

## Implementation
Implement [the new
protocol](https://github.com/typedb/typedb-protocol/pull/224).

The two operations are implemented as two separate services running in
`tokio` tasks, similar to transaction services.

### Database export 
This operation is simple. After establishing a stream, we just export
the database's schema (the same operation as schema retrieval) and then
send a number of items containing the header, the database's data
(encoded concepts), and data checksums at the end.

### Database import
This operation is more tricky. It is executed in steps, containing a
couple of "states":
1. The database's name and schema are expected. Without it, we can't
continue. Ignoring this step leads to an error, signaling that the
client is probably implemented incorrectly.
2. After the schema is received, it is executed and committed as is, to
check that the provided schema is correct. Otherwise, a user error is
returned, so they can rewrite the schema and try again.
3. After the schema is persisted, it is relaxed. This contains: a)
substituting default cardinalities and card/key annotations with
`@card(0..)` for all capabilities in the schema b) making all attributes
independent (`@independent`) so they are not cleaned up between
transactions without owners if they are not yet received c) making all
relations independent (a new system `system independent` property, not
exposed to users) so they are not cleaned up between transactions
without role players. All errors at this stage are considered internal
errors and bugs.
4. After the schema is prepared, we receive items, decode them into
concepts, and persist in the database. There is a transaction buffer
size we use to execute commits from time to time, reducing memory
consumption and final commit time. This optimization requires 3b and 3c.
One of the expected and required items is checksums. Another optional
item is the header (we don't really use it outside of logs, so it's not
required. Not sure if we need to force it).
5. After the `Done` message is received, signaling that the stream of
items is completed, we perform the final data commit and unrelax the
schema (undo 3).
6. If everything is good, a verification `Done` response is sent.
At any point, an error can be returned. To make it possible, the
protocol introduces streaming from the server, although it remains
silent until the end unless there are errors in the provided messages.

While being imported, databases are not accessible through
`database_manager`: they are owned by `database_importer`. If a server
crashes, the uncompleted databases will be cleaned up on the next
bootup.

To avoid overflowing memory, `InstanceIDMapping` uses `SpilloverCache`,
a new component combining `HashMap` and RocksDB to spill over the
excessive data which is too much to fit in memory.