Remove "no open transaction" errors from HTTP transaction close endpoint (#7487)
## Product change and motivation
Remove the last possible and very rare case of the "no open transaction"
error after calling `/transactions/:transaction-id/close` when the
transaction is closed in parallel without answering the current request.

## Implementation
Transaction requests create a channel:
```
let (result_sender, result_receiver) = oneshot::channel();
```
and then await the result in `result_receiver`. It can return an error
if `result_sender` is dropped -> if `TransactionService`'s
`request_stream` is dropped -> if `TransactionService` is dropped -> if
`TransactionService` is closed -> transaction is closed or committed.