Simplify PartialCostHash and beam search implementation (#7495)
## Product change and motivation
We simplify the hash used to detect redundant states during the plan-search, loosening the definition of equivalent plans for more aggressive pruning.

## Implementation
We update the `PartialCostHash` to have a stronger definition of 
Since we want to choose the plan with the lowest cost amongst all plans corresponding to the same state, we convert `new_plans_heap` to a min_heap so the first one we see is the best one.