Add HTTP authentication action diagnostics counts (#7511)
## Product change and motivation
Add diagnostics reporting for authentication actions performed through
the HTTP endpoint.

Previously, due to its excessive usage for every request and low value,
we ignored this method. However, it's the most efficient metric to
understand the usage of the HTTP endpoint of the server, so it is a
valuable piece of usage metrics for us.

## Implementation
We additionally cleanup some unused code and comments.