Improve intersection cost estimate, granular query plan cache eviction (#7513)
## Product change and motivation

We improve the cost estimate of an intersection by factoring in the work
required to produce each element of the iterators involved. Previously
the cost estimate assumed that said cost was limited to fetching a
single key on disk, however when a post-filter is applied to a
constraint iterator, we may have to fetch multiple keys to produce one
that is accepted by all filters.

Additionally, the query plan cache now takes into account the exact
types used in the query to make eviction decisions on a plan-by-plan
basis.

## Implementation

Other minor improvements:
- short circuit during an intersection if any iterator is empty,
- reduce allocations when iterating extensions during beam search.