Only seek on iterators that are behind in merge (#7526)
## Product change and motivation

Fix a crash when during a seek on a `KMergeBy`, which is merging
multiple sorted tuple iterators, `seek` is called on iterators that are
already ahead of the seek target, an invalid operation. It is valid to
`seek` the iterator as a whole, as the merge may still be behind the
seek point even if some constituent iterators are ahead.

## Implementation