Implement optionality in read queries (#7508)
## Product change and motivation

We implement a huge new feature: optionality in read queries! This means
you can now do partial matches.

### Basics 
In the following query, all `person` instances will be returned, and if
any name exists, the name will be added to the answer. However, _not_
having an answer will _not_ cause that person instance to be eliminated:
```
# Find any person, and if they have a name add that into the answer
match 
  $p isa person;
  try { $p has name $name; };
```
The answers will conceptually conform to a type like `[ $p=person,
$name=Option<name>]`.

Optionals only return when the _full_ pattern inside the optional
matches:
```
# Return any people, and use the ages only of people with name "john", else do not match anything additional
match 
  $p isa person;
  try { $p has name "john"; $p has age $age; };
```

Multiple `try` clauses are matched to the maximum degree at
simultaneously for each non-optional answer component:
```
# If every person can have at most 1 name and 1 age: return every person, plus their name if it exists, plus their age if it exists
match
  $p isa person;
  try { $p has name $name; };
  try { $p has age $age; };
```
Note that this will produce 1 answer per person, not a permutation of
each person, with and without their names and ages. If there are higher
cardinalities of name or age, you get permutations of each possible name
and age filled but neither filled.

Optional variables cannot be re-used across sibling clauses due to
ambiguous outcomes:
```
### ILLEGAL - optional $name cannot be re-used across sibling try clauses
match 
  $p isa person;
  try { $p has name $name; }; 
  try { $x isa person, has name $name; };
```

### Nesting and Pipelining

Optionals can also not be nested within disjunctions (`or`) or negations
(`not`) - in negations, these operations would be no-ops, and
disjunctions are currently limited to only returning variables that are
guaranteed to be returned from any branch, ie. they cannot currently
return optional values at all.

Optional variables _can_ be used across stages of a query pipeline. Any
pattern in a query using an optional variable that is not set will
"fail" immediately (return 0 answers). However, by nesting a pattern
containing an unset optional variable inside try` clause, this behaviour
is absorbed since only the `try` clause will fail:

```
match
  try { $name isa name "John"; };
match
  try { $p isa person, has name $name; };

```
This will return exactly `[$name=None, $p=None ]` or `[$name=Some(name),
$p=Some(person) ]` or `[$name=Some(name), $p=None]`

In contrast, by using the optional `$name` _outside_ any optional scope,
the following query will eliminate any answers that don't have `$name`
set and only return answers containing `[$name=name, $person=person]`
```
match
  try { $name isa name "John"; };
match
  $p isa person, has name $name;
```


## Implementation

We've made several architectural changes to how we think about the
structures present in the `compiler` package.

### Centralising "scoping semantics" into the BlockContext and Patterns

Attempted to move towards the `BlockContext` having enough information
on its own in order to have all the information required to determine
variable visibility, binding, and legality (eg. no sharing variables
between sibling optionals!)

This was not entirely completed, though we made two changes that
contribute towards this.

* `Transparency` to `ScopeType`
`Transparency` was used to help determine if a variable was "visible"
through a nested scope or not - which depended on whether it was a
Disjunction, Negation, or Conjunction. However, this was not general
enough to consider Optionals as well.

Instead, we now just record the type of scope directly (Disjunction,
Negation, Conjunction, or Optional) into the BlockContext, and let the
user of the Context figure out the visibility rules based on this. In
theory, since the BlockContext now has full information about scope
parent-child combinations, and also the type of the scope, we are almost
at the state at which we could determine all semantic information about
variables, visibility, and binding requirements. What is missing is the
information to determine if a variable must be an input (eg. bound
elsewhere) or if it can be bound in the pattern scope.

This information is currently only visible directly in Patterns, which
takes us the the second change:

* `VariableDependency` to `VariableBindingMode`
Dependency used to indicate if a variable was producing, required input,
or referencing. This strategy did not really extend to optionals and was
generally confusing. It's been replaced with a trait that produces a
"Binding mode" for each variable in a Pattern - including recursively
defined or referenced variables in child patterns! These are 'context
free', in that information about the scopes that are parents to a scope
are irrelevant.

The Binding mode provides full information about each variable having to
be `PreBound` (eg same as required input before - expression arguments,
function arguments, etc.), `AlwaysBinding` (say bound in every branch of
a disjunction, or if this is a Conjunction), `OptionallyBinding` (for
optionals that have local variables that are also visible to the
parents), and `LocallyBindingInChild` (for variables in less than all
branches of a disjunction, or variables locally used only in a
negation).

A variable's binding mode can be different when queried at any level of
a pattern - in a `Negation`, which just contains a `Conjunction`, all
variables are probably `AlwaysBinding`. However, at the Negation's level
they are probably mostly `LocallyBindingInChild`, etc.

To help use this information, we have a new trait on all `NestedPattern`
which is called `Pattern`, that exposes the raw context-free
`VariableBindingMode` for each variable, but also a set of APIs that are
not context-free and utilise the BlockContext to see what parents are
available from the parents - for example,
`required_inputs(blockContext)`,
`visible_binding_variables(blockContext)`, and
`named_visible_binding_variables(blockContext)`.

### Simplifications to Planning

We eliminate some redundancy in the planner functions by noticing that
scoped pattern planning graphs are build recursively, and already have
all the required information about what is a required input to a
particular scope's planner. There previously was confusion generated by
setting variables as "inputs" when they were already given as "shared
variables" during construction as well.

We also streamling naming throughout the planner, and settle on calling
the representations of "sub-plans" in a planning graph "xxxVertex",
which is how they are used and perceived by the parent planner.

#### Futuer Improvements
One future improvement of the planner should be achieved by noting that
the "Graph" terminology is not very explicit. There is the idea of a
"Vertex", but no idea of an "edge" - if we had adopted this naming, we'd
notice that there is redundancy in a Graph, which stores both
VariableVertex -> PatternVertex edges and PatternVertex ->
VariableVertex edges, but at the same time each Pattern can also
generate the set of `VariableVertex` again.

A simpler model would also explicitly write out the bipartite graph:
```
Graph {
  variables:
  patterns:
  variables_to_patterns:
  patterns_to_variables:
}
```
And then possible decide to either keep `patterns_to_variables` as the
only way to traverse from patterns to variables, or only keep a
`PatternVertex.variables()` as the only way to traverse from patterns to
variables.

---------

Co-authored-by: Krishnan Govindraj <4918770+krishnangovindraj@users.noreply.github.com>